+++ src/utils/string-utils.ts
